Generate mock arrays with minimum number of items
The previous code always generated 1 item in the array.  If the property
has a minItems value that is >1 then the mock will fail response validation.

This change fixes it by generating the minimum number of items in the
array to match minItems.  If minItems is not defined, then it will
continue to generate 1 item as before.

Test plan:
* Added unit tests to cover minItems set and not set
* Used gulp to run lint and all unit tests and ensured everything passes.
* Used coverage report to verify all new lines were tested at least once.