Handle nested defaults in body parameters
Body parameters (`in` is `"body"`) are always grouped in a single Object,
and as such there is only the option to set a `default` for the whole
body (or none).

This change looks deeper into the object for `default`s to support more
flexible configuration of expected parameters in `body`.  This brings
the body support closer to e.g. `path` which already supports multiple
parameters.

This change will look for defaults parameter values for all `object`
type parameters.  This is also recursive, so if an `object` has a
parameter that is an `object` then this code will look for defaults in
that nested object.

Note: this has only been implemented for Swagger 2.0 as I am not
familiar with Swagger 1.2, so felt it best to leave 1.2 as-is.

Test Plan:
- Test that `gulp` runs and passes, ensuring:
-- new code passes lint rules
-- existing unit tests all run and pass
- Add new tests for testing the new features, and confirm all pass.
- Check the code coverage, and verify that (almost) all new lines have
  been covered.
-- The only new lines not being covered is the exclusion of Swagger 1.2
   schemas, which has only been checked visually.